deleted the old component and renamed the new one
