--- modules/shared/components/molecules/ImgWithInfo copy/IImgWithInfoDemo.d.ts
@@ -1,20 +1,0 @@
-declare namespace IImgWithInfoDemo {
-  export interface IProps extends IImage {
-    classes?: string;
-    children?: ReactElement;
-    avatarSize?:
-      | 'extra-large'
-      | 'extra-small'
-      | 'large'
-      | 'medium'
-      | 'small'
-      | 'super-large';
-    variant?: 'anonymous' | 'filled' | 'notFilled';
-    profilePic?: string;
-    onAvatarClickHandler?: () => void;
-    ImageVariant?: 'avatar' | 'icon';
-    isHidden?: boolean;
-  }
-}
-
-export { IImgWithInfoDemo };
