--- modules/shared/components/molecules/ImgWithInfo copy/ImgWithInfoDemo.module.css
@@ -1,15 +1,0 @@
-.outer-wrapper {
-  @apply flex w-min;
-}
-
-.user-wrapper {
-  @apply max-w-13xl flex flex-col justify-between ml-4 whitespace-nowrap;
-}
-
-.title {
-  @apply text-sm font-normal block whitespace-nowrap overflow-hidden overflow-ellipsis;
-}
-
-.subTitle {
-  @apply text-xs font-light	text-dark-grey block;
-}
