+++ modules/shared/components/molecules/ImgWithInfo/ImgWithInfo.tsx
@@ -1,44 +1,54 @@
 import React from 'react';
-import type { ReactElement, FC } from 'react';
+import type { ReactElement } from 'react';
+import classNames from 'classnames';
+import styles from './ImgWithInfo.module.css';
 import type { IImgWithInfo } from './IImgWithInfo';
 import Avatar from '../../atoms/Avatar/Avatar';
-import styles from './ImgWithInfo.module.css';
+// this line case an error when we ran the test, the error : cannot find module or it is corresponding type
+// import { handleImgWithInfoVariant } from '@modules/shared/logic/imgWithInfoVariant/imgWithInfoVariant';
 import { handleImgWithInfoVariant } from '../../../logic/imgWithInfoVariant/imgWithInfoVariant';
 
-const ImgWithInfo: FC<IImgWithInfo.IProps> = ({
-  isHidden,
-  profilePic,
-  title = '',
-  subTitle,
-  description,
-  variant,
+const ImgWithInfo = ({
+  classes,
   children,
-}): ReactElement => {
-  const username = title.length !== 0 ? title : 'Anonymous';
+}: IImgWithInfo.IProps): ReactElement => {
+  const boxClasses = classNames(styles.box, classes);
+  return <div className={boxClasses}>{children}</div>;
+};
 
+const Image = ({
+  classes,
+  avatarSize = 'medium',
+  ImageVariant,
+  profilePic,
+  onAvatarClickHandler,
+  isHidden = false,
+  children,
+}: IImgWithInfo.IProps): ReactElement => {
+  const ImageClasses = classNames(styles['header-wrapper'], classes);
   return (
-    <div className={styles['outer-wrapper']}>
-      {variant === 'avatar' && (
-        <Avatar
-          size="medium"
-          variant={handleImgWithInfoVariant(isHidden, profilePic)}
-          profilePic={profilePic}
-        />
-      )}
-      {variant === 'icon' && children}
-      <div className={styles['user-wrapper']}>
-        <span className={styles.title} data-testid="title">
-          {username}
-          {isHidden && title && (
-            <span className="ml-2 text-grey text-xs">(anonymous)</span>
-          )}
-        </span>
-        <span title={description} className={styles.subTitle}>
-          {subTitle}
-        </span>
+    <div className={ImageClasses}>
+      <div className={styles.header}>
+        {ImageVariant === 'avatar' && (
+          <Avatar
+            size={avatarSize}
+            variant={handleImgWithInfoVariant(isHidden, profilePic)}
+            profilePic={profilePic}
+            onClick={onAvatarClickHandler}
+          />
+        )}
       </div>
+      {ImageVariant === 'icon' && children}
     </div>
   );
 };
+
+const Info = ({ classes, children }: IImgWithInfo.IProps): ReactElement => {
+  const bodyClasses = classNames(classes);
+  return <div className={bodyClasses}>{children}</div>;
+};
+
+ImgWithInfo.Image = Image;
+ImgWithInfo.Info = Info;
 
 export default ImgWithInfo;
